web-animations/animation-node/animation-node-player.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>AnimationNode player attribute tests</title>
<meta name="assert" content="The player with which this animation node is associated, if any">
<link rel="help" href="http://dev.w3.org/fxtf/web-animations/#widl-AnimationNode-player">
<link rel="help" href="http://dev.w3.org/fxtf/web-animations/#dfn-associated-with-a-player">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
test(function() {
    var node = newAnimation(createDiv(this));

    assert_equals(node.player, null, 'AnimationNode.player should be null');
}, 'AnimationNode.player is null if animation node is not associated with a player');

test(function() {
    var node = new AnimationGroup([]);

    assert_equals(node.player, null, 'AnimationNode.player should be null');
}, 'AnimationNode.player is null if animation node is not associated with a player. ' +
    'Test AnimationGroup');

test(function() {
    var node = newAnimation(createDiv(this));
    var player = document.timeline.play(node);

    assert_equals(node.player, player, 'AnimationNode.player ' +
        'should return player associated with this node');
}, 'AnimationNode.player returns player associated with this node');

test(function() {
    var node = new AnimationGroup([]);
    var player = document.timeline.play(node);

    assert_equals(node.player, player, 'AnimationNode.player ' +
        'should return player associated with this node');
}, 'AnimationNode.player returns player associated with this node. ' +
    'Test AnimationGroup');

test(function() {
    var node1 = newAnimation(createDiv(this));
    var node2 = newAnimation(createDiv(this));
    var node3 = newAnimation(createDiv(this));
    var node4 = new AnimationGroup([node3]);
    var animationGroup = new AnimationGroup([node1, node2, node4]);
    var player = document.timeline.play(animationGroup);

    assert_equals(node1.player, player, 'AnimationNode.player should return the player of its ancestor');
    assert_equals(node2.player, player, 'AnimationNode.player should return the player of its ancestor');
    assert_equals(node3.player, player, 'AnimationNode.player should return the player of its ancestor');
    assert_equals(node4.player, player, 'AnimationNode.player should return the player of its ancestor');
    assert_equals(animationGroup.player, player, 'AnimationNode.player should return player ' +
        'directly associated with this node');
}, 'AnimationNode.player returns the player of its ancestor, ' +
    'if it is not directly associated wtih the player');

test(function() {
    var node1 = newAnimation(createDiv(this));
    var node2 = newAnimation(createDiv(this));
    var animationGroup = new AnimationGroup([node1, node2]);
    var player = document.timeline.play(animationGroup);
    animationGroup.player.currentTime = ANIMATION_END_TIME/2;

    assert_equals(window.getComputedStyle(node1.target).width, 100,
        'Nodes in the hierarchy should be controlled by parent group player');
    assert_equals(window.getComputedStyle(node2.target).width, 100,
        'Nodes in the hierarchy should be controlled by parent group player');
}, 'Nodes in the hierarchy are controlled by parent group player');
</script>
</body>
